Rename, tweaks
