+++ tests/integration/status-list-module.test.js
@@ -13,7 +13,7 @@ import { createNewDockDID } from "../../src/utils/did";
 import { OneOfPolicy } from "../../src/utils/revocation";
 import { registerNewDIDUsingPair } from "./helpers";
 import getKeyDoc from "../../src/utils/vc/helpers";
-import StatusList2021Credential from "../../src/status-list-credential/status-list-2021-credential";
+import StatusList2021Credential from "../../src/status-list-credential/status-list2021-credential";
 
 const expectEqualCreds = (cred1, cred2) => {
   expect(cred1).toEqual(cred2);
